fix: Ensuring inheritance
The call to `ensure_inheritance` can cause a stack overflow when
used in combination with the `enumerize` gem and a standard inherited
model (not STI - no type column defined on the table).

https://github.com/gocardless/statesman/pull/373#discussion_r465877454

There's no need to continue with the `ensure_inheritance` when using a
standard inherited model; the statesman methods are already available
on the subclass.